feat: dual mode should also support mock
