+++ mcpserver/core/base.py
@@ -2,7 +2,11 @@ import collections
 import json
 import time
 
+from resource_secretary.providers import discover_providers
+from resource_secretary.providers.mock import discover_mock_providers
+
 import mcpserver.utils as utils
+from mcpserver.logger import logger
 
 
 class WorkerBase:
@@ -11,6 +15,17 @@ class WorkerBase:
     ask secretary. We provide it here so that a hub can use it to generate
     its dual mode (acting as worker AND hub.)
     """
+
+    def init_providers(self, mock=False):
+        """
+        Probe the local system on startup. E.g., "we found spack, flux, etc."
+        These can be faux (mock) or real discovered providers
+        """
+        logger.info("📡 Probing local system for resource providers...")
+        if mock:
+            self.catalog = discover_mock_providers(self.worker_id, choice=mock)
+        else:
+            self.catalog = discover_providers()
 
     def register_agent_tools(self):
         """
